% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demo.R
\name{maggie_says}
\alias{maggie_says}
\title{maggie_says}
\usage{
maggie_says(says)
}
\arguments{
\item{says}{the string that you want maggie to say}
}
\value{
a printed statement that maggie says
}
\description{
maggie_says
}
\examples{
maggie_says('hi')
}
